Renamed userReducer to userSlicer
